refact(watch): move Reason into build_loop
Reason doesn’t have any business in watch, literally everything but
FilesChanged is a build_loop concept.

This is made obvious by the fact that the module only ever uses the
`FilesChanged` constructor and then `if let`s (ugh) on it in the
tests.
